refactor: api 폴더 위치 변경 및 상대 경로 적용
- 모듈 불러오기 .js suffix 제거
- api.js -> request.js로 파일 네이밍 변경
