+++ src/js/components/header/Header.js
@@ -1,8 +1,8 @@
-import { removeLocalStorageItem } from '../../utils/storage.js';
-import { headerTemplate } from './headerTemplate.js';
-import { $ } from '../../utils/dom.js';
-import { showSnackbar } from '../../utils/snackbar.js';
-import { SELECTOR, PATH, STORAGE, SUCCESS_MESSAGE } from '../../constants.js';
+import { removeLocalStorageItem } from '../../utils/storage';
+import { headerTemplate } from './headerTemplate';
+import { $ } from '../../utils/dom';
+import { showSnackbar } from '../../utils/snackbar';
+import { SELECTOR, PATH, STORAGE, SUCCESS_MESSAGE } from '../../constants';
 
 class Header {
   #props;
@@ -15,18 +15,18 @@ class Header {
 
   init(isLoggedIn) {
     this.#isLoggedIn = isLoggedIn;
-    this.initDOM();
-    this._selectDOM();
-    this._bindMenuEvent();
+    this._initDOM();
   }
 
-  initDOM() {
+  _initDOM() {
     this.$target = $(SELECTOR.HEADER);
     this.$target.innerHTML = headerTemplate(this.#isLoggedIn);
+    this.$menu = $(SELECTOR.MENU);
+    this._bindEvent();
   }
 
-  _selectDOM() {
-    this.$menu = $(SELECTOR.MENU);
+  _bindEvent() {
+    this._bindMenuEvent();
   }
 
   _bindMenuEvent() {
