+++ src/js/components/station/stationTemplate.js
@@ -39,18 +39,17 @@ export const stationsTemplate = stations => {
 export const stationTemplate = (id, { name }) => {
   return `
   <li class="station-list-item d-flex items-center py-2" data-station-id=${id}>
-  <span class="station-list-item-name w-100 pl-2">${name}</span>
-  <button
-  type="button"
-  class="modify-button modal-trigger-btn bg-gray-50 text-gray-500 text-sm mr-1"
->
-  수정
-</button>
-<button type="button" class="delete-button bg-gray-50 text-gray-500">
-  삭제
-</button>
-</li>
-`;
+    <span class="station-list-item-name w-100 pl-2">${name}</span>
+    <button
+    type="button"
+    class="modify-button modal-trigger-btn bg-gray-50 text-gray-500 text-sm mr-1"
+    >
+      수정
+    </button>
+    <button type="button" class="delete-button bg-gray-50 text-gray-500">
+      삭제
+    </button>
+  </li>`;
 };
 
 export const modalTemplate = () => {
